Fixed some warnings

+++ scripts/Binary.py
@@ -11,7 +11,7 @@ from skimage import img_as_ubyte
 def im_threshold(quant):
     quant = cv2.cvtColor(quant, cv2.COLOR_BGR2LAB)[:,:,0]
 
-    binary = np.ones_like(quant) * 255
+    binary = np.ones_like(quant, dtype=np.uint8) * 255
     binary[quant == np.min(quant)] = 0
  
     return img_as_ubyte(binary)
@@ -21,7 +21,7 @@ def im_threshold(quant):
 # clusters: cluster pixels
 # imshape: image shape
 def bin_from_clusters(clusters, imshape):
-    bim = np.ones(shape=imshape)
+    bim = np.ones(shape=imshape, dtype=np.uint8)
 
     for cluster in clusters:
         bim[cluster[:, 0], cluster[:, 1]] = 0
